% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{box_cox}
\alias{box_cox}
\title{Apply / invert a Box-Cox power transformation}
\usage{
box_cox(x, lambda)
}
\arguments{
\item{x}{Numeric vector of strictly positive values.}

\item{lambda}{Numeric scalar. Power parameter.}
}
\value{
A numeric vector of transformed values.
}
\description{
The Box-Cox transformation (Box & Cox, 1964) is
\deqn{y(\lambda) = \begin{cases}
  (x^\lambda - 1) / \lambda & \lambda \neq 0 \\
  \log(x) & \lambda = 0.
\end{cases}}
}
\details{
For \code{lambda = 0} this returns \code{log(x)}; for \code{lambda = 1} it returns
\code{x - 1} (no shape change). Use \code{\link[=shewhart_box_cox]{shewhart_box_cox()}} to estimate
\code{lambda} from the data via profile log-likelihood.
}
\examples{
box_cox(1:10, lambda = 0)      # equivalent to log(1:10)
box_cox(1:10, lambda = 0.5)

}
\references{
Box, G. E. P., & Cox, D. R. (1964). An Analysis of Transformations.
\emph{Journal of the Royal Statistical Society, Series B}, 26(2), 211-252.
\doi{10.1111/j.2517-6161.1964.tb00553.x}
}
\seealso{
\code{\link[=shewhart_box_cox]{shewhart_box_cox()}} to estimate lambda from data.
}
