% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{color_hue}
\alias{color_hue}
\title{Generate a qualitative HCL palette}
\usage{
color_hue(n)
}
\arguments{
\item{n}{Integer (>= 1). Number of colours to generate.}
}
\value{
A character vector of \code{n} hex colour strings.
}
\description{
Internal palette used as a default when a number of phases or
subgroups need distinct colours. For most plots, ggplot2's default
palette is fine; this is provided for backward compatibility and
quick prototyping.
}
\examples{
color_hue(5)

}
