% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvd_recife}
\alias{cvd_recife}
\title{COVID-19 daily mortality, Recife, Brazil, 2020}
\format{
A tibble with 279 rows and 3 columns:
\describe{
\item{date}{Date of the bulletin.}
\item{new_deaths}{Integer count of new deaths reported that day.}
\item{.t}{Integer row index (1..N), useful as a continuous
predictor for \code{\link[=shewhart_regression]{shewhart_regression()}}.}
}
}
\source{
Castlab (Universidade Federal de Pernambuco) compiled
the original series from the Brazilian Ministry of Health daily
bulletins. See \url{https://covid.saude.gov.br/}.
}
\usage{
cvd_recife
}
\description{
Daily count of new COVID-19 deaths officially recorded in Recife
(capital of Pernambuco state, Brazil) between 28 March 2020 and
31 December 2020. Distributed with v0.1.x as
\code{inst/extdata/recife_2020_covid19.rds} and preserved in v1.0.0
as a vignette case study illustrating the regression-based control
chart on non-stationary epidemiological counts.
}
\examples{
\donttest{
fit <- shewhart_regression(cvd_recife,
                           value = new_deaths, index = .t,
                           model = "loglog")
ggplot2::autoplot(fit)
}
}
\seealso{
\code{\link[=shewhart_regression]{shewhart_regression()}},
\code{vignette("covid-recife", package = "shewhartr")}.
}
\keyword{datasets}
