% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs-tests.R
\name{shewhart_rules_available}
\alias{shewhart_rules_available}
\title{List available runs rules}
\usage{
shewhart_rules_available()
}
\value{
A tibble with columns \code{rule} (the key) and \code{description}.
}
\description{
List available runs rules
}
\examples{
shewhart_rules_available()
}
