% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.shewhart_chart}
\alias{summary.shewhart_chart}
\title{Compact tibble-like summary}
\usage{
\method{summary}{shewhart_chart}(object, ...)
}
\arguments{
\item{object}{A \link{shewhart_chart} object.}

\item{...}{Currently unused.}
}
\value{
A list with elements \code{limits} and \code{violations}.
}
\description{
Compact tibble-like summary
}
\examples{
fit <- shewhart_i_mr(data.frame(y = rnorm(50)), value = y)
summary(fit)

}
