% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{inv_box_cox}
\alias{inv_box_cox}
\title{Inverse Box-Cox transformation}
\usage{
inv_box_cox(x, lambda)
}
\arguments{
\item{x}{Numeric vector of strictly positive values.}

\item{lambda}{Numeric scalar. Power parameter.}
}
\value{
A numeric vector on the original scale.
}
\description{
Inverse Box-Cox transformation
}
