% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.shewhart_chart}
\alias{tidy.shewhart_chart}
\title{Tidy the control limits of a Shewhart chart}
\usage{
\method{tidy}{shewhart_chart}(x, ...)
}
\arguments{
\item{x}{A \link{shewhart_chart} object.}

\item{...}{Currently unused.}
}
\value{
A tibble with at least columns \code{chart}, \code{line}, \code{value}.
}
\description{
Returns a tibble of the chart's control limits in tall format. Each
row corresponds to one line of one chart panel (CL / UCL / LCL).
}
\examples{
fit <- shewhart_i_mr(data.frame(y = rnorm(50)), value = y)
broom::tidy(fit)

}
